+++ Common/Integrations/GenericModConfigMenu/IGenericModConfigMenuApi.cs
@@ -15,5 +15,6 @@ namespace Common.Integrations.GenericModConfigMenu
         void RegisterChoiceOption(IManifest mod, string optionName, string optionDesc, Func<string> optionGet, Action<string> optionSet, string[] choices);
         void RegisterClampedOption(IManifest mod, string optionName, string optionDesc, Func<int> optionGet, Action<int> optionSet, int min, int max);
         void RegisterClampedOption(IManifest mod, string optionName, string optionDesc, Func<float> optionGet, Action<float> optionSet, float min, float max);
+        void SetDefaultIngameOptinValue(IManifest mod, bool optedIn);
     }
 }