+++ HorseFluteAnywhere/release-notes.md
@@ -2,7 +2,9 @@
 
 # Release notes
 ## Upcoming release
-* When configuring the mod through Generic Mod Config Menu, you can now set [multi-key bindings](https://stardewvalleywiki.com/Modding:Player_Guide/Key_Bindings).
+* When using [Generic Mod Config Menu](https://www.nexusmods.com/stardewvalley/mods/5098), you can now...
+  * configure the mod after loading the save file;
+  * set [multi-key bindings](https://stardewvalleywiki.com/Modding:Player_Guide/Key_Bindings).
 * Fixed compatibility with [unofficial 64-bit mode](https://stardewvalleywiki.com/Modding:Migrate_to_64-bit_on_Windows).
 
 ## 1.1.3